// src/app/Models/Vendor Models/vendorReg.ts
export class vendorReg {
  active: string;
  address: string;
  ccEmailIds: string;
  contactNumber: string;
  createdBy: string;
  createdByName: string;
  createdDateTime: string;
  email: string;
  id: string;
  isVendor: string;
  location: string;
  message: string;
  newPassword: string;
  password: string;
  status: string;
  updatedBy: string;
  updatedByName: string;
  updatedDateTime: string;
  vendorId: string;
  vendorName: string;

  cpass: string;
  contactName: string;
  vendorNameSearch: string;
}